+++ R/scorecard.R
@@ -35,9 +35,9 @@ ab = function(points0=600, odds0=1/19, pdo=50) {
 #' @param odds0 Target odds, default 1/19. Odds = p/(1-p).
 #' @param pdo Points to Double the Odds, default 50.
 #' @param basepoints_eq0 Logical, default is FALSE. If it is TRUE, the basepoints will equally distribute to each variable.
-#' @return scorecard
+#' @return A scorecard dataframe
 #'
-#' @seealso \code{\link{scorecard_ply}}
+#' @seealso \code{\link{scorecard2}} \code{\link{scorecard_ply}}
 #'
 #' @examples
 #' \dontrun{
@@ -70,6 +70,12 @@ ab = function(points0=600, odds0=1/19, pdo=50) {
 #' # scorecard
 #' # Example I # creat a scorecard
 #' card = scorecard(bins, m)
+#' card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
+#'   x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
+#'    "purpose", "credit.amount", "savings.account.and.bonds",
+#'    "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
+#'    "personal.status.and.sex", "other.debtors.or.guarantors", "property",
+#'    "age.in.years", "other.installment.plans", "housing"))
 #'
 #' # credit score
 #' # Example I # only total score
@@ -80,7 +86,6 @@ ab = function(points0=600, odds0=1/19, pdo=50) {
 #' }
 #' @import data.table
 #' @export
-#'
 scorecard = function(bins, model, points0=600, odds0=1/19, pdo=50, basepoints_eq0=FALSE) {
   # global variables or functions
   variable = var_woe = Estimate = points = woe = NULL
@@ -124,20 +129,23 @@ scorecard = function(bins, model, points0=600, odds0=1/19, pdo=50, basepoints_eq
   return(card)
 }
 
-
-#' Score Transformation
+#' Creating a Scorecard
 #'
-#' \code{scorecard_ply} calculates credit score using the results from \code{scorecard}.
+#' \code{scorecard} creates a scorecard based on the results from \code{woebin} and \code{glm}.
 #'
-#' @param dt Original data
-#' @param card Scorecard generated from \code{scorecard}.
-#' @param only_total_score  Logical, default is TRUE. If it is TRUE, then the output includes only total credit score; Otherwise, if it is FALSE, the output includes both total and each variable's credit score.
-#' @param print_step A non-negative integer. Default is 1. If print_step>0, print variable names by each print_step-th iteration. If print_step=0, no message is print.
-#' @param replace_blank_na Logical. Replace blank values with NA. Default is TRUE. This argument should be the same with \code{woebin}'s.
+#' @param dt A data frame with both x (predictor/feature) and y (response/label) variables.
+#' @param y Name of y variable.
+#' @param x Name of x variables. If it is NULL, then all variables in bins are used. Default is NULL.
+#' @param bins Binning information generated from \code{woebin} function.
+#' @param points0 Target points, default 600.
+#' @param odds0 Target odds, default 1/19. Odds = p/(1-p).
+#' @param pdo Points to Double the Odds, default 50.
+#' @param basepoints_eq0 Logical, default is FALSE. If it is TRUE, the basepoints will equally distribute to each variable.
+#' @param positive Value of positive class, default "bad|1".
+#' @param ... Additional parameters.
+#' @return A scorecard dataframe
 #'
-#' @return Credit score
-#'
-#' @seealso \code{\link{scorecard}}
+#' @seealso \code{\link{scorecard}} \code{\link{scorecard_ply}}
 #'
 #' @examples
 #' \dontrun{
@@ -170,6 +178,104 @@ scorecard = function(bins, model, points0=600, odds0=1/19, pdo=50, basepoints_eq
 #' # scorecard
 #' # Example I # creat a scorecard
 #' card = scorecard(bins, m)
+#' card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
+#'   x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
+#'    "purpose", "credit.amount", "savings.account.and.bonds",
+#'    "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
+#'    "personal.status.and.sex", "other.debtors.or.guarantors", "property",
+#'    "age.in.years", "other.installment.plans", "housing"))
+#'
+#' # credit score
+#' # Example I # only total score
+#' score1 = scorecard_ply(dt, card)
+#'
+#' # Example II # credit score for both total and each variable
+#' score2 = scorecard_ply(dt, card, only_total_score = F)
+#' }
+#' @import data.table
+#' @export
+scorecard2 = function(bins, dt, y, x=NULL, points0=600, odds0=1/19, pdo=50, basepoints_eq0=FALSE, positive='bad|1', ...) {
+  variable = NULL
+
+  setDT(dt)
+  # bins # if (is.list(bins)) rbindlist(bins)
+  if (!is.data.table(bins)) {
+    if (is.data.frame(bins)) {
+      bins = setDT(bins)
+    } else {
+      bins = rbindlist(bins, fill = TRUE)
+    }
+  }
+
+  # check x and y
+  dt = check_y(dt, y, positive)
+  # x
+  x_bins = bins[, unique(variable)]
+  if (is.null(x)) x = x_bins
+  x = x_variable(dt,y,x)
+
+  # dt to woe values
+  dt_woe = do.call(woebin_ply, args = c(list(dt=dt, bins=bins, print_info=FALSE), list(...)))
+
+  # model
+  model = glm(
+    as.formula(paste(y, "~ .")), family = "binomial",
+    data = dt_woe[,c(paste0(x,"_woe"),y), with=F])
+
+  return(scorecard(bins = bins, model = model, points0 = points0, odds0 = odds0, pdo = pdo, basepoints_eq0 = basepoints_eq0))
+}
+
+#' Score Transformation
+#'
+#' \code{scorecard_ply} calculates credit score using the results from \code{scorecard}.
+#'
+#' @param dt Original data
+#' @param card Scorecard generated from \code{scorecard}.
+#' @param only_total_score  Logical, default is TRUE. If it is TRUE, then the output includes only total credit score; Otherwise, if it is FALSE, the output includes both total and each variable's credit score.
+#' @param print_step A non-negative integer. Default is 1. If print_step>0, print variable names by each print_step-th iteration. If print_step=0, no message is print.
+#' @param replace_blank_na Logical. Replace blank values with NA. Default is TRUE. This argument should be the same with \code{woebin}'s.
+#'
+#' @return A dataframe in score values
+#'
+#' @seealso \code{\link{scorecard}} \code{\link{scorecard2}}
+#'
+#' @examples
+#' \dontrun{
+#' # load germancredit data
+#' data("germancredit")
+#'
+#' # filter variable via missing rate, iv, identical value rate
+#' dt_sel = var_filter(germancredit, "creditability")
+#'
+#' # woe binning ------
+#' bins = woebin(dt_sel, "creditability")
+#' dt_woe = woebin_ply(dt_sel, bins)
+#'
+#' # glm ------
+#' m = glm(creditability ~ ., family = binomial(), data = dt_woe)
+#' # summary(m)
+#'
+#' # Select a formula-based model by AIC
+#' m_step = step(m, direction="both", trace=FALSE)
+#' m = eval(m_step$call)
+#' # summary(m)
+#'
+#' # predicted proability
+#' # dt_pred = predict(m, type='response', dt_woe)
+#'
+#' # performace
+#' # ks & roc plot
+#' # perf_eva(dt_woe$creditability, dt_pred)
+#'
+#' # scorecard
+#' # Example I # creat a scorecard
+#' card = scorecard(bins, m)
+#' card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
+#'   x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
+#'    "purpose", "credit.amount", "savings.account.and.bonds",
+#'    "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
+#'    "personal.status.and.sex", "other.debtors.or.guarantors", "property",
+#'    "age.in.years", "other.installment.plans", "housing"))
 #'
 #' # credit score
 #' # Example I # only total score